+++ ChangeLog.md
@@ -45,6 +45,7 @@
 - BUGFIX: #4267 Path now shows familiar backslashes on Windows for Add a database and About dialog.
 - BUGFIX: #4314 Fixed missing OpenSSL license exemptions.
 - BUGFIX: #4433 Fixed loading translations under Linux.
+- BUGFIX: #4447 Fixed AUTOINCREMENT generation of DDL statement for table-level PRIMARY KEY.
 
 ### 3.3.3
 - CHANGE: #4011 SQLite updated to 3.35.4. SQLite3MultipleCiphers updated to 1.2.4 (SQLite 3.35.4). This enables math functions from 3.35. SQLCipher updated to 4.4.3, which is only SQLite 3.34.1, so no math functions for SQLCipher.